csp_documentation/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
env
requests==2.31.0
PyPDF2==3.0.1
groq==0.4.2
pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.9
Office365-REST-Python-Client==2.5.0 